Complete folder structure with running server
