+++ src/schema/builder.ts
@@ -7,6 +7,17 @@ export const builder = new SchemaBuilder<{
   Context: {
     db: PrismaClient;
   };
+  Objects: {
+    Task: Task;
+  };
 }>({});
+
+interface Task {
+  id: string;
+  title: string;
+  completed: boolean;
+  createdAt: Date;
+  updatedAt: Date;
+}
 
 builder.addScalarType("Date", DateResolver, {});
