+++ src/schema/index.ts
@@ -1,4 +1,5 @@
 import { builder } from "./builder";
 import "./task";
+import "./query";
 
 export const schema = builder.toSchema({});
